Clang-tidy & cppcheck

+++ src/qcommon/q_string.h
@@ -24,7 +24,7 @@ char *Q_strupr( char *s1 );
 char *Q_strrchr( const char* string, int c );
 
 // buffer size safe library replacements
-void Q_strncpyz( char *dest, const char *src, int destsize );
+void Q_strncpyz( char *dest, const char *src, size_t destsize );
 void Q_strcat( char *dest, int size, const char *src );
 
 const char *Q_stristr( const char *s, const char *find);
